feat: Add QuestionFactory
